requirements.txt
pymupdf
sentence-transformers
chromadb
google-genai
python-dotenv
numpy
rank_bm25
